Move feedback buttons from thread messages to editor component
This commit relocates the feedback functionality (thumbs up/down) from the last assistant message to the message editor component. This provides a more consistent UI where feedback can be given at any time rather than only for the last assistant message.

Changes include:
- Removing feedback UI from ActiveThread component
- Adding feedback buttons to MessageEditor
- Removing unneeded workspace dependency from ActiveThread
- Simplifying message rendering in ActiveThread
