+++ crates/assistant2/src/assistant_panel.rs
@@ -171,7 +171,6 @@ impl AssistantPanel {
                 thread.clone(),
                 thread_store.clone(),
                 language_registry.clone(),
-                workspace.clone(),
                 window,
                 cx,
             )
@@ -245,7 +244,6 @@ impl AssistantPanel {
                 thread.clone(),
                 self.thread_store.clone(),
                 self.language_registry.clone(),
-                self.workspace.clone(),
                 window,
                 cx,
             )
@@ -379,7 +377,6 @@ impl AssistantPanel {
                         thread.clone(),
                         this.thread_store.clone(),
                         this.language_registry.clone(),
-                        this.workspace.clone(),
                         window,
                         cx,
                     )
